move to predicates module
