+++ kernel/src/expressions/mod.rs
@@ -12,8 +12,6 @@ pub use self::column_names::{
 pub use self::scalars::{ArrayData, Scalar, StructData};
 use crate::DataType;
 
-pub mod parquet_stats_skipping;
-
 mod column_names;
 mod scalars;
